10.26 14:21 Bug / Fixed / ToDo.
Bug :
- GuiMember, GuiNoCreated press 'E' key cause gui exit.
Fixed :
- Message display for two player interact with same block.
- Chat system can't process system command send from client.
To Do :
- Language file localization.
- GuiAokChat for better interface.
- Admin in-game gui terminal for setting global value.
- Lord level definition.
- Economy system.

This commit changed some stuff such as canceled the originial chat
packet handler instead use vanilla one, add few lang file statement for
localization, catch a bug from our testing server which announced in
terminal that world saving data might cause a NullPointerException, but
it still don't know why and we are tracing.
Until now, this project is still working on it, and we finally got a
server for test, it would be welcome to everyone who wants to help us
improove this project. if you do so, you can conntect me through Issues
page of this project! :wink: